trialscout-backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.6.4
pydantic-settings==2.2.1
python-dotenv==1.0.1
httpx==0.27.0
sqlalchemy==2.0.36
alembic==1.14.0